Switch to Prettier 2.0 for our Javascript code
The main change is that now parentheses are always added around the
formal parameters of an arrow function.

The default of trailingComma also has been changed from "none" to "es5",
but we set this option to "always" anyway, so this does not affect us
(and "always" is likely to become the future default of Prettier anyway).