Fix #563: do not fail container mode if bind-mounted files exist
Overlayfs does not support being mounted over a file, only over
directories. If there was a bind mount of a file, creating the container
failed if overlayfs was to be used for that part of the file-system
hierarchy. Now we just create a read-only mount of the file
(the only thing that is possible anyway).