Uneject react-scripts from our HTML table
This effectively reverts 40e43ade2a8ec2069bf48d69d75345a1d1705ef3.

The advantage is that we are not locked with the configuration and
scripts that we currently have in the repo but get updates.
For example, the current set of files is already incompatible with newer
versions of react-dev-utils (cf. dde1ad9a453fbbc2a6899e3311fcfe2b5e2a4a80).

Typically ejecting react-scripts is a one-way operation,
but we can manually revert it by doing this:
- Install react-scripts. When we ejected we had version 3.0.1 and it is
  a good idea to keep the version as similar as possible for now,
  but we need some features from 3.1.0 for being able to locally
  override our eslint and jest configs, so we pick 3.1.0.
  This actually leads to a downgrade of many of our development
  dependencies, but we can update them later again.
- Remove development dependencies that are not allowed as direct
  dependencies if react-scripts are used (the latter has these as
  transitive dependencies anyway).
- Remove all files from config/ and scripts/ that were not changed.
- Update package.json to remove jest and scripts config
  that is already provided by react-scripts.
- We have several customizations to webpack.config.js.
  To be able to keep them, install package react-app-rewired
  and move the whole webpack config to config-overrides.js.
  This is a hacky workaround and can create problems,
  but it works for now and we can do further cleanup later.
- Let scripts/start.js (which we also have customized) only contain our
  own changes and then delegate to the standard script.

The only changes from the ejection that are still remaining
is the babel config in package.json (otherwise nothing works)
and the added dependencies (which we will remove in a separate step).

This change actually provokes a changed vendors.min.js bundle,
probably due to the downgraded versions of webpack etc.
There is also a wrong eslint warning due to the older eslint version.
But we can ignore both for now as we are going to upgrade all packages
next anyway.