Make it easier to select link-text in tables
Some browsers (e.g. Firefox) do interpret
click+select on 'blocks' that are links as click+drag.
This makes it impossible to select the text in those links
if there is no space around the links to start clicking..
Make this possible by changing links in tables from 'block' to 'inline-block'.
This way, the links do not take up the full cell and it is
possible to select the text with click+select.
Add a small margin to the links so that there is always
a little space to start the click.

Downside: Not the full cell is a link anymore,
but only a reasonable space around the actual link text.