Fix table-generator: compute scores based on property files
So far this was done based on the name of the property,
but this was never clearly defined.
The behavior also caused wrong tables in case of SV-COMP property files
with uncommon names, because while table-generator still applied
the property files correctly to the rows, it did not correctly compute
scores (this even happened in one of our integration tests).

Note that because of this change, for tasks with task-definition files,
not only the task-definition files but also the property files
need to be available during table generation in order for scores to be computed.

To do this, we now open and parse each task-definition file several
times (once for every run set plus once afterwards).
We could probably cache that somehow,
but we will add more task metadata to result files soon,
which will avoid the need for reading task-definition files
in table-generator completely (except for old results).