Start working on making project REUSE compliant
For example, add LICENSES folder with all relevant licenses
and improve documentation on licenses.
We now also declare all licenses of code that is contained in our
releases in setup.py.

We use a .reuse/dep5 file to declare licenses of files where adding
a copyright header would not be possible or worth it:
- generated files used in tests
- generated JS bundles
- property files (cannot be changed)
- several small files in debian/ where it is unclear whether comments
  are supported