Ensure that licenses are added to bundled files
New versions of wheel and setuptools are needed for this.
This works for the source tar and the wheel,
but unfortunately not the egg.
I guess we can just drop producing the latter.

Now that we ship license texts, we can also include the custom license
in the metadata.
Previously that did not make much sense because users could not look it
up.