Fix buttons below plots in the HTML table (#578)
The recently added function `setParam` (for adding plot state to the URL
such that it can be handled by hash routing) calls a popstate event, which 
refreshes the state of the React object according to the parameters in 
the URL. As Reacts `setState` function is async, setting the state 
afterwards again (depending on the current state) can result in unwanted 
behavior, such as the state being overwritten with the old values again 
or switching back to the initial values. This second setting of the 
state is not necessary any more and creates bugs, so it should be 
removed.

Fixes #577.