Fix incorrect cgroup fallback from 03e282d8
When BenchExec is started and we do not have permissions for the current
cgroups of the process, we want to try
system.slice/benchexec-cgroup.service as fallback,
because the recommended way of installing BenchExec creates this cgroup.
However, instead of trying system.slice/benchexec-cgroup.service,
we tried to find a cgroup with the name below the current cgroup.
This does not make sense, because the fallback cgroup is only ever
created at the root of the cgroup hierarchy.
So the fallback would only work if the current cgroup happened to be the
root cgroup.