Make systemd config files no longer depend on JoinControllers option
We used that option to make systemd create hierarchies for all
controllers that we need, instead of only for those that systemd uses by
default.
Unfortunately, with systemd 240 the option was removed.
This affects for example Ubuntu 20.04.

So now we change benchexec-cgroup.service such that it supports several
independent hierarchies (one for each controller)
and creates the system.slice/benchexec-cgroup.service cgroup in each
hierarchy if systemd hasn't done it.

This works in principle in the same way as before,
just that the docs of systemd don't give us as many guarantees as
before: because we set Delegate=true in the unit we could be sure that
cgroups that systemd created for us won't be touched.
Now we have to hope that systemd doesn't interfere with the cgroups that
we created in the other hierarchies.

The benchexec-cgroup.service definition still supports the old case
(with JoinControllers), it does not matter how many hierarchies exist.
The same is true for BenchExec itself (and has always been).
But our Debian package and our installation instructions won't use
JoinControllers anymore, even on systems where it would be supported,
to keep things simple.

Furthermore, this change also improves two other issues:
- There is no need anymore to update the initramfs and reboot.
  Installation can be done online.

- Some software (specifically snapd) did not work on systems with
  JoinControllers, this is now no longer a problem, of course.

Fixes #580