When neither the current nor the fallback cgroup work, use the latter
Obviously, this will only affect the error message that is printed
later on to the user.
Furthermore, we do this only if the fallback cgroup actually exists.
The effect is that on systems where the fallback cgroup exists
but we have insufficient permissions, we can produce a more precise
error message, because the user is likely intending to use the fallback
cgroup (which is the recommended way on systemd machines).
All other cases are not changed.