Improve error messages on failure to use cgroups
We now try to distinguish different failure cases and give more specific
hints on what the user should do:
- cgroups missing completely vs. just permission problems
- whether permission problems could be fixed by adding user
  to a set of groups or not
- systemd running or not
- Debian/Ubuntu or something else

Furthermore, we collect all cgroup problems
and give a single final error message about the most severe problem.
This means that if several required cgroups are unusable,
the user is notified about all of them,
and that we can print a single command as workaround
for permission problems instead of one per subsystem.