Make the timestamps in filenames more precise and better readable
So far, benchexec produces result files with names like
"...2020-06-09_1234...".
The missing seconds cause issues if rapidly repeating benchmarks,
and the time part of it is hard to read.
So we change the format to "...2020-06-09_12:34:56".

This means that "benchexec --startTime" now also needs to take seconds
(but this parameter is rarely used anyway).

The same change is made for tables produced by table-generator
(for those cases were table-generator adds a timestamp).

Fixes #283.