CPAchecker tool-info: don't add the status twice if the msg is the same
This change has been made in particular because of the portfolio algorithm:
For a successful analysis, the result gets added twice to the log, i.e.
once for the successful subanalysis, and once for the main analysis itself
(which 'copies' the result from the subanalyis)

When using table-generator, this has led to results being shown in
the table e.g. as 'true (true)', or 'false(unreach-call) (false(unreach-call))',
where for the former the pink color was then shown, i.e. the same as if an
error or timeout would have happened.