Improve error message if empty task-definition file is read
If an empty task-definition file is read,
pyyaml does not raise an exception, but returns `None`.

This was not handled and resulted in an AttributeError.
Print a nice error message instead.

Easiest way to reproduce:

```
touch doc/empty.yml
bin/benchexec doc/benchmark-example-true.xml
```

This will try to read the doc/empty.yml and provoke
the described error.