Create directories for hidden mountpoints in container
A hidden mountpoint is a mountpoint in a directory where we do not have
permissions. Because of the missing permissions, we cannot mount
something over this mountpoint, so we go upwards and mount an empty
directory over the next accessible parent directory.
However, the kernel will still show the original mountpoint in its list
of mounts, but now there is not even a directory for it (because a
parent was replaced with an empty directory).
This confuses tools that expect that a mountpoint always exists as a
directory (an example for this is containerexec itself).
So now we ensure that an (empty) directory exists for such a hidden
mountpoint.

A common case where this situation occurs it the mountpoint
/sys/kernel/debug/tracing, which is inaccessible because of the
permissions of /sys/kernel/debug.

This makes nesting containerexec work without the need for special
parameters to the inner containerexec.
So we can also avoid special-casing /sys in tests of nested containers.