Add 'context' and 'machine_model' to task-def YML
'context' contains a dict that can contain further
information about the context that a task definition
is defined in, i.e., context that is expected by the tasks
that are derived from the task definition.

'machine_model' is a key in the dict of 'context',
and specifies the machine model that is assumed for
the properties on the input files of the task.