Add and use an adapter from BaseTool to BaseTool2 for tool-info modules
It will be automatically applied if the tool-info module provides an
instance of BaseTool such that the rest of the code can rely on the API
of BaseTool2 without the need for changes to all tool-info modules.

Note that there is a technical backwards incompatibility here:
Previously it was not checked whether the class of a tool-info module
actually inherited from BaseTool as long as it provided the necessary
methods, now we do check that it is an instance of either BaseTool or
BaseTool2.
However, it was always documented that the Tool class of a tool-info
module needs to inherit from BaseTool,
so there is no incompatibility with tool-info modules that conform to
the documentation.

For now we intend to keep this adapter (and support for old tool-info
modules) until at least the next major version of BenchExec.