Refactor internal handling of resource limits in benchexec
Now that we have a proper class ResourceLimits, let's use it internally
as well instead of the previous dict.

This also changes how time limits are represented.
Previously, benchexec internally stored a time limit and an optional
soft time limit. If no soft time limit was given, the time limit was
used in its place.
Externally (e.g., in the benchmark definition) it always supported
providing a time limit and an optional hard time limit, though,
with the semantics of the external time limit being those of the
internal soft time limit, and the semantics of the external hard time
limit being those of the internal time limit.

Now we just store time limit and hard time limit internally
with the same semantics as exposed to the user.
If only one of them is given, the other is set to the same value.