Fix benchexec bug in combination of <withoutfile> and working directory
If a tool-info module specifies a different working directory than the
current directory and <withoutfile> was used in the benchmark definition
to specify tasks, the task identifier was accidentally treated as a path
and made relative to the working directory, although this does not make
sense.
Now it is passed through unchanged.