Reduce file I/O for writing result files in benchexec
During benchmarking, benchexec writes intermediate result files
such that nothing is lost in case something bad happens.
While the XML result file was written at most every 60s
in order to reduce I/O, the text result file was rewritten completely
after each run.
This means that the I/O produced for this text file scaled quadratically
with the number of runs and could reach several hundred gigabytes
for 40000 runs.

We could rewrite the file more rarely like we do for the XML file,
but for the text file there is a better approach.
During benchmarking, we now write the file incrementally, one line per run.
After the run set, we truncate the file and rewrite it properly once.

This means that during benchmarking one can notice two differences
in the text file's content:
- no lines for runs that are not yet started
- results are ordered as the runs finish instead of proper order
Arguably, at least the former item is actually an advantage as it allows
to watch the file with tail during benchmarking.
After the benchmarking, no difference in content will exist compared to
the previous behavior.