Slightly change definition of isTimeout flag for tool-info modules
Previously it was set if either the termination reason indicated a
timeout or if the consumed time was higher than the limit.
The reason for having the latter check at all was that the termination
reason was not fully reliable, but it should be reliable now.
Furthermore, the only reason why isTimeout is passed to the tool-info
module is such that the latter can distinguish between what seems like a
crash due to being killed after the time limit and a real crash,
but for this use case it is only important to know whether benchexec
killed the tool, which is reflected in the termination reason.
So checking the limits for this is not necessary and we remove this.