Fix bug from dc84cc89347a2f87cfe53c527e830e7b13383082
The new replacement for rlimits[TIMELIMIT] is rlimits.cputime_hard.
rlimits.cputime replaces rlimits[SOFTTIMELIMIT].

It also makes sense to use the hard time limit in awsexecutor,
because that limit defines how many resources a run is allowed to use,
where as the main time limit is used only for deciding whether
the result counts.