fix unpacking of String into a list of chars.
There was a regression with BenchExec 3.3 in CPAchecker.
See our Buildbot jobs for the Java analysis:
https://buildbot.sosy-lab.org/cpachecker/#/builders/14/builds/853

The problem is the auto-unpacking of lists into tuples.
This behaviour is bad for us, as Strings are handled as a list of chars.

Example:
`tuple([1,2,3]) --> (1,2,3)` # no more list
We better use the following here (the trailing comma is important!):
`([4,5,6],) --> ([4,5,6],) # tuple with one element, which is a list