Change heuristic for detecting wrong includesfile tags in benchexec
So far we checked whether a given includes file contains {,
which would likely be a source-code file instead of a set file,
indicating confusion between <include> and <includesfile>.
However, this is not helpful when the same confusion occurs
and task-definition files are used.

Thus we now simply abort if nothing in a set file matches an existing
file.