Move package metadata from setup.py to setup.cfg
A declarative declaration of metadata is of course much nicer
as inside some Python code.
Furthermore, because we already require a newer setuptools version,
we can make use of features like automatic extraction of version and
description from files instead of having to do this manually.

The resulting metadata of sdist archives and wheels is identical
except for a small syntactic change in the keywords list.