avoid unix-style path separator.
This is some kind of usability improvement,
because that path might be printed for the user.
The script seems to work without this char anyway.

There are more code locations where we use unix-style path separators.
None of them seems to cause any problem on Windows (as far as tested)
and most of them never are never shown towards the user.