remove equivalence check for property files during parsing task definitions.
ossible reason for the problem:
ProbablyThe method `os.path.samefile` does not work reliably on Windows when using network storage.

Using VirtualBox with shared directories,
Python seems to be not able to distinguish reliably between different property files.
I do not know the exact reason, but the following benchmark definition causes a weird behaviour:
(requires CPAchecker r35613 and a current BenchExec version)
  $> scripts/benchmark.py @webcloud test/test-sets/integration-bam.xml

The benchmark is expected to have 1847 tasks,
and Linux shows that number when submitting the tasks.
Windows however finds for the same benchmark 2070 tasks,
because it ignores some invalid property file matches.
I have even seen other numbers, between the 1847 and 2070.
The bug seems to depend on the provided inodes for those files.
I do not know, when files need to be identical, whether this is a bug
in our scripts or in Python (3.9.0 for Win64) or in VirtualBox 5.2.42.
The change in these lines fixes the bug, however it removes the equivalence check.
I do not know a case where the equivalence check would have helped.