Replace floats with Decimal in tablegenerator
Decimal has many advantages compared to floats such as support for
correct rounding and the ability to precisely represent decimal numbers.

Decimal was already used for statistics calculations, now we also use it
when rounding and formatting values.

This commit contains the minimal amount of changes necessary to make
this work, without any further fixes and simplifications.

Due to the increased precision of Decimal, there are now also
differences in the output of our integration tests,
many of which are actually improvements.
To keep the commit size small, these will be tackled in subsequent commits,
so the table-generator integration tests are expected to fail now.