Adjust HTML expected results due to improvements in arithmetic precision
Cf. the previous two commits

This commit changes only the expected "raw" values
that are contained in the JSON and used by the statistics code
for filtering etc., but are not shown to the user.
Some changes occur in plots, but also with invisibly small differences.
The changes for the raw values are the same as in the previous commit for
the CSV output, i.e., obvious improvements.

The integration tests of table-generator still fail because there are
also changes to the rounded and formatted values
(which are shown to the user) due to the switch to Decimal.
These will be handled in the next commit
because some of them are actually regressions
(due to previous imprecision hiding another bug).