Fix output of very small values in tooltips of HTML table
Our formatting code incorrectly estimated the number of required digits
(because str(value) switches to scientific notation for small values,
and thus the resulting string is shorter than what we actually need)
and thus incorrectly truncated small values to zero.

This only affected the tooltips with statistics in HTML tables,
because other values are formatted differently.