Add our own formatting for Decimal instances
The builtin formatting always falls back to exponential notation
for very small values, which we do not want
(added test case for this).
Using string formats such as "%f" would avoid exponential notation
but has the drawback that we need to specify a precision.

Writing such a formatting was actually not that difficult
because Decimal.as_tuple() exists and provides exactly the required
information.
This will probably also allow us to simplify the other formatting code
(that handles rounding and alignment).