Fix output format of scaled non-rounded values in table-generator
For values that were not scaled and not rounded (e.g., in columns of
type count) we always printed the original value with all its digits
in CSV files, including trailing zeros.
For scaled values, however, we omitted trailing zeros.
Example: Input value "0.1230" would be printed as "0.1230" if not
scaled, but as "1.23" after scaling by 10.

Now we change that such that the original number of digits is used.
In the example, "1.230" is now printed.
This was actually desired initially but was not implemented.
Due to the switch to Decimal we can now do this easily.

This affects only CSV files and the "raw" values of HTML tables
(used internally).
So this does not actually have any visible effect, but is still good for
consistency.

To keep the commit size small, the necessary test adjustments will be
done in separate commits,
so the table-generator integration tests are expected to fail now.