Improve formatting of numbers in table-generator
We can use our new utility for printing Decimals,
and thus no longer need to estimate how many decimal digits we need.
There were still some cases where the previous logic led to incorrect
truncation of very small values (unit test is also added).

Due to this commit and the previous one, the resulting formatted number
is already almost exactly as desired, so we can reduce the postprocessing
steps that manipulate the number as a string.