Remove weird unused case in tablegenerator
We trimmed the leading 0 of values like 0.1 in format_number
either if the value is to be printed in a cell of the HTML tables,
or in CSV tables with alignment.
Such CSV tables are not even produced by tablegenerator,
and it is not clear why alignment should have an effect on whether to
trim leading zeros.