Fix logic for rounding tooltip_stochastic values in tablegenerator
There are cases (columns of type count without specified number of
significant digits) where we do not have a good idea on how to round
(because columns of type count do not need rounding,
but the tooltip_stochastic values for these columns do need to be
rounded), so we have always done some arbitrary rounding.

However, the same rounding was also applied for columns of type measure
without specified significant digits, where it makes more sense to apply
the same logic for tooltip_stochastic as for the actual values in the
table cells.

Note that this does not currently have any visible effect because
tooltip_stochastic is actually not used currently (which is wrong).