Fix handling of avg and stdev values in tooltips of HTML tables
The values min, max, and median can always simply be handled with the
same rounding as the actual values in the table cells,
because after all all three values do appear in some table cells.
The same rounding can also be applied for sum, because it has the same
properties.
However, avg and stdev are different, most notably because these are not
guaranteed to be integer values even if the whole column contains only
integer values. So we need to force some rounding here.
This was done in the past, but not applied for the new kind of HTML
tables.

This affects only columns of type count (only integer values),
because for other columns tooltip_stochastic will be handled in the same
way as the other types (cf. previous commit).

There are two main visible effects (only in the tooltips of HTML tables):
- For values that happen to be integral,
  we still print digits after the decimal separator.
- For non-integral values, we avoid output like
  79876269.42500396282727840969 and round nicely.