Fix rounding of some statistics values in tablegenerator
We check the number of existing significant digits and make sure to
never output more significant digits in the result than there are in the
input. However, for statistics values this included converting Decimal
instances to strings with str(), which does not always do what we want.
For example, for small values it uses scientific notation and it does
some normalization for 0, so we lost the number of significant digits.