Fix alignment of statistics values in summary table of HTML tables
Since the introduction of the new kind of HTML tables,
the values in the summary table were rendered with the same alignment
as the values in the actual table.
However, for the summary table we often need less whitespace for alignment,
because the values in it often have less digits after the decimal
separator (the summary table only shows sums, so values tend to be larger).

Now we compute the alignment for the summary table based only on the
values in the summary table.

This was originally added in 4f534c74f101fc55445a50234e416847fc9603a4
but was not used for the new kind of tables and the old code was removed
in 2c4b5872873e4b31fa6415a54740526308b823c1.

In the past, alignment information (specifically the number of digits
after the decimal separator) was also used in the actual rounding code,
not only for alignment, so changing Column instances could also have an
effect on formatting of tooltips, for example.
However, since the recent improvements the number of decimal digits is
really only used for alignment, so this does not affect anything else.