Improve handling of initial table state
For consistency, we require the initial / because it is a regular part
of the hash part of our URLs, but we give a helpful error message if it
is missing.
For the same reason, we require the value in the JSON
to be exactly the URL part after the hash.
Also improve the documentation and given an example.