Change corner case of statistics computation for HTML tables
For computing the sum of values, we returned None for an empty list,
but 0 for any non-empty list even if it contained only None values
and not a single numeric value.
Now we return None as well for a list of only None values,
because this still means "no values available" like an empty list.
In the tables, None will be rendered as a dash.
It is more useful for the reader to be able to distinguish between
"no numeric values available" and "at least one real value available"
then between "this is a statistic over zero table rows" and
"this is a statistic over at least one table row",
because these two cases can be distinguished anyway simply by looking
at how many rows per category exist.