Fix passing of "raw" values to HTML tables
In contrast to what the name suggests, these values were sometimes
rounded, specifically if the user had explicitly requested a certain
number of significant digits.
However, this can lead to undesired effects if the JavaScript code
uses these values for computing statistics,
so now we always pass the non-rounded value as "raw" value.
All JavaScript code expects and handles non-rounded values there anyway.
This does not affect most tables at all,
and even for those where the value was previously rounded
this does not affect the values that are printed in the table,
it only affects what the table internally uses when filtering etc.