Fix printing of trailing zero digits in stdev of HTML tables
These were missing because we started calculating the standard deviation
with the value "0" and then when printing did not add more decimal
digits than present.
Now we simply add many trailing zeros to the initial value of stdev.

This affects only the stev value in tooltips of HTML tables
and only adds trailing zeros.