Fix rounding of certain cases of the value 0 in HTML tables
For some reason, the code did not take the intended number of
significant digits into account for the value 0,
differently from all other values.
This was probably a leftover from then the formatting of the value 0
was completely separate from the formatting of the other values.
Now we fix this and make it consistent for all values.

This affects only how many trailing digits of the number 0 are printed,
and in practice affects only the stdev value in tooltips
(which were often just "0" instead of something like "0.000").