Update documentation of requirements, particularly the kernel version
According to the release notes of systemd version 238,
Linux 4.14 or newer is recommended for the memory cgroup,
so we can simplify our recommendations.

Furthermore, we noticed that some of our hard-coded constants
of the Linux ABI have different values on rare platforms (e.g., MIPS),
so BenchExec is not expected to work there.
It is probably not worth the effort to implement support for all
platforms, so let's just document this.
We also add a C program that can be used to test our current assumptions
of the Linux ABI (x86 and ARM were tested with this and work).