Fix incorrect escaping of table-generator cmdline that is printed
The path to table-generator was not escaped at all,
so even whitespace was a problem,
and the paths to the XML files were wrong in case they contained
a ' character.
Now we use a proper escaping method,
which also avoids quotes in the common case where they are not needed.