Improve handling of empty result files in table-generator
So far we returned a run set with zero columns,
which led to weird effects in the HTML tables,
for example that the run set was completely invisible
(because zero columns wide) in all tables.
Thus we add at least a dummy status columns (without entries) now.

The same fallback was actually also already used for result files
that had some results, but none of the results had any columns.

This would have also fixed #733.