Add close() method to tool-info modules
The main reason for this is that we did already add a noop close method
(at least without container mode) to tool-info modules,
and overwrote any method that a tool-info module might have added.
This could in theory lead to the tool-info module calling our close
method instead of the expected one.
Now a method with this name is documented to exist and never
overwritten, so this is avoided.

Furthermore, it helps type checks and typing consistency in BenchExec,
because we no longer call close on BaseTool2 instances
that was not guaranteed to exist in the class definition.