Update React-Table to v7
This updates the libary for the tables in the Summary and Table tab to 
the newest version. Since this update is a major rewrite of the lib that 
leaves the whole frontend to the end-user, large amounts of the code had 
to be changed. Additionally, uniqid was explicitly added to the 
development dependencies of the project, as this became necessary with 
the update.