Clarify and somewhat strengthen guarantees about called methods
The order of method calls was already guaranteed,
but strictly speaking that those methods are always called was not
guaranteed. But this was the original intention (and the actual behavior
of BenchExec) because the explicitly stated goal was to allow tool-info
modules to safely store some cache in executable() and similar methods.
And of course this only works if executable() is always called.