Fix computation of confidence intervals
Previously, the ending points of the regression line were artificially 
added to the data which is passed to the function for the confidence 
intervals as well, which is why 2 was subtracted from the number of data 
points. Since this is not the case any more, this subtraction can be 
removed.