Introduce threshold for number of data points of regression lines
Previously, regression plots with a large number of data points may 
require a huge amount of resources or even crash. To reduce the costs of 
the plot computation, a maximum number of data points for the regression 
and confidence interval lines is introduced at a value of 10.000. If 
there are more data points, the lines are only plotted at points in 
fixed intervals such that a number of 10.000 is never exceeded. In 
practice, this is not visually noticeable in the plot.