Fix a few problems with the statistics table
These occurred since the switch to react-table v7.
Among them are a missing horizontal scrollbar,
truncation of table width,
and broken stickyness of the left-most column.

Fixing them is done by adjusting the DOM structure and the CSS
such that more definitions are applied to both the main table
and the statistics table.

Not all problems are fixed yet,
for example resizing columns is still broken.

Thanks to @skanav for reporting this.