Improve CSS and DOM for clickable elements in table headers
Header cells that are not able to trigger sorting no longer get the
necessary definitions.
We apply CSS to indicate clickable elements only to those elements
and not to all header cells with the need to revert it for some.
On the other hand, this allows to easily apply the definition for
clickable elements also to the statistics table,
where they were missing (part of #742).