Introduce limit for how many log messages about result files are created
With --debug and container mode, BenchExec logs about every single file
that was created by the tool, matches the specified result-files
pattern, and is copied to the output directory.
This can lead to arbitrarily large log output, which is not useful
and often unexpected for users, because typically the runexec log
is relatively short and its size does not depend on what the tool does.
So to avoid dumping unexpectedly large logs
onto users and consuming tools, we just log the first 1000 files.
But we can even improve the usefulness of the log
by also logging how many files were copied in total,
which was not available before.