If loading logs from local tables fails show hint for local HTTP server
Part of #797

Browsers are more and more restricting the possibility to access other
local files via Ajax from a local HTML document.
We need this in order to show logs if the user opens an HTML table
that is stored on their local disk.
A workaround is for the user to start a local HTTP server
that serves both the table as well as the logs and open the table
via that server instead of a file:// URL.
This should be easy for most BenchExec users because Python includes
such an HTTP server and BenchExec users typically have Python installed,
but most do not know about this.
Furthermore, that basic HTTP server does not support directory listings,
so they cannot navigate to their table, they need to manually create the
correct URL (relative to the document root of the started HTTP server).

So in order to make this usable we explain what to do in the help
message that occurs if loading a log from a local table fails.
We can even generate the full necessary command line
(with the longest common prefix of both the current table and the
 current log as document root) for copy-pasting
and a link that will point to the table once the HTTP server was started.
By keeping the hash part in the URL the table will even reload with the
same settings.
This should work in all cases except if the user does not have Python or
if port 8000 is already used.