Improve help message for starting a HTTP server for local tables
So far we used a trick of prefixing "//." to the server's base directory
to make it work with both Linux ("/home/...") and Windows ("/C:/...")
paths. However, this does not work for network-share paths on Windows
("//host/..."). Furthermore, if the table and the logs are on different
partitions on Windows, there does not exist any common base directory
that we can specify.
Now we handle these two cases and avoid using confusing (but working)
"//." prefix.