Fix refactoring from 8b2d46c1
The test version of the duplicated code that was removed
did not actually do the same thing.
The real code used the computed statistics only for adding a row of
filtered statistics, whereas the test code overwrote all statistics
with the values computed by the statistics worker.
This was done in order to have tests that check whether the statistics
worker returns exactly the same values as the original statistics from
table-generator.
So since then this test didn't actually do anything meaningful.

Because there was much more duplicate code before the refactoring,
we do not revert it but simply add an option to the code.