Revert cdad9e67: Filtered statistics again completely replace what is shown
Previously we had switched to showing the statistics over the filtered
tasks in a separate row in the statistics table
and kept all standard statistics visible.
However, this is inconsistent with how the HTML tables treat filters
in other areas, e.g., in plots.
The discussion in #125 showed
that simply replacing all visible statistics with the statistics
over the filtered data is preferred, so we do this now.
In order to make this clear for users who know the previous behavior,
we add " of selected rows" as a suffix of the statistics name.

If filters are present, we also hide all statistics that we cannot
compute on-the-fly, such as local summary and score.