Do not show unfiltered statistics as fallback for filtered statistics
When we load the table with already applied filters
(e.g., because they are encoded in the URL),
the previous behavior was to first show the default statistics
(those that are computed by table-generator and are over all data)
and then switch to the filtered statistics once their values
are computed by the JS statistics worker.
This is of course confusing and misleading.
So now we show a placeholder message in such cases instead.